refactor: special dates component
